Move function AppController.handleReferenceClick to verseListController
